+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- Added fake metrics gatherer.
+- Gauge widget.
+- Main term application.
+- Fake metrics gatherer.
 
 [unreleased]: https://github.com/slok/go-http-metrics/compare/v0.x.0...HEAD
 [0.x.0]: https://github.com/slok/meterm/releases/tag/v0.x.0